feat: inline discover customization
