+++ src/components/MovieDetails/index.tsx
@@ -462,7 +462,7 @@ const MovieDetails = ({ movie }: MovieDetailsProps) => {
                   key={`keyword-id-${keyword.id}`}
                 >
                   <a className="mb-2 mr-2 inline-flex last:mr-0">
-                    <Tag content={keyword.name} />
+                    <Tag>{keyword.name}</Tag>
                   </a>
                 </Link>
               ))}
