+++ src/styles/globals.css
@@ -68,7 +68,7 @@
   }
 
   .slider-header {
-    @apply relative mt-6 mb-4 flex;
+    @apply relative mb-4 flex;
   }
 
   .slider-title {
